utils: add htlc & swap related helpers
