+++ swap.go
@@ -4,11 +4,10 @@ import (
 	"context"
 	"time"
 
-	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/lightninglabs/lndclient"
 	"github.com/lightninglabs/loop/loopdb"
 	"github.com/lightninglabs/loop/swap"
-	"github.com/lightningnetwork/lnd/input"
+	"github.com/lightninglabs/loop/utils"
 	"github.com/lightningnetwork/lnd/lntypes"
 )
 
@@ -50,59 +49,10 @@ func newSwapKit(hash lntypes.Hash, swapType swap.Type, cfg *swapConfig,
 	}
 }
 
-// GetHtlcScriptVersion returns the correct HTLC script version for the passed
-// protocol version.
-func GetHtlcScriptVersion(
-	protocolVersion loopdb.ProtocolVersion) swap.ScriptVersion {
-
-	// If the swap was initiated before we had our v3 script, use v2.
-	if protocolVersion < loopdb.ProtocolVersionHtlcV3 ||
-		protocolVersion == loopdb.ProtocolVersionUnrecorded {
-
-		return swap.HtlcV2
-	}
-
-	return swap.HtlcV3
-}
-
 // IsTaproot returns true if the swap referenced by the passed swap contract
 // uses the v3 (taproot) htlc.
 func IsTaprootSwap(swapContract *loopdb.SwapContract) bool {
-	return GetHtlcScriptVersion(swapContract.ProtocolVersion) == swap.HtlcV3
-}
-
-// GetHtlc composes and returns the on-chain swap script.
-func GetHtlc(hash lntypes.Hash, contract *loopdb.SwapContract,
-	chainParams *chaincfg.Params) (*swap.Htlc, error) {
-
-	switch GetHtlcScriptVersion(contract.ProtocolVersion) {
-	case swap.HtlcV2:
-		return swap.NewHtlcV2(
-			contract.CltvExpiry, contract.HtlcKeys.SenderScriptKey,
-			contract.HtlcKeys.ReceiverScriptKey, hash,
-			chainParams,
-		)
-
-	case swap.HtlcV3:
-		// Swaps that implement the new MuSig2 protocol will be expected
-		// to use the 1.0RC2 MuSig2 key derivation scheme.
-		muSig2Version := input.MuSig2Version040
-		if contract.ProtocolVersion >= loopdb.ProtocolVersionMuSig2 {
-			muSig2Version = input.MuSig2Version100RC2
-		}
-
-		return swap.NewHtlcV3(
-			muSig2Version,
-			contract.CltvExpiry,
-			contract.HtlcKeys.SenderInternalPubKey,
-			contract.HtlcKeys.ReceiverInternalPubKey,
-			contract.HtlcKeys.SenderScriptKey,
-			contract.HtlcKeys.ReceiverScriptKey,
-			hash, chainParams,
-		)
-	}
-
-	return nil, swap.ErrInvalidScriptVersion
+	return utils.GetHtlcScriptVersion(swapContract.ProtocolVersion) == swap.HtlcV3
 }
 
 // swapInfo constructs and returns a filled SwapInfo from
